requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0

# Environment Variables
python-dotenv==1.0.0

# AI and Graph
langgraph==0.0.20
ollama==0.1.0

# MCP
mcp==0.1.0

# Database
psycopg2-binary==2.9.9

# HTTP Client
httpx==0.25.0
